Move backend-agnostic dump_mir_items into kani_middle
